Unfriend ReSolve from its HiOp interface.
