+++ src/LinAlg/CMakeLists.txt
@@ -100,6 +100,9 @@ set(hiopLinAlg_PARDISO_SRC
 set(hiopLinAlg_CUSOLVER_LU_SRC
   hiopLinSolverSparseCUSOLVER.cpp
   IterativeRefinement.cpp
+  RefactorizationSolver.cpp
+  MatrixCsr.cpp
+  KrylovSolverKernels.cu
 )
 
 set(hiopLinAlg_CUSOLVER_CHOL_SRC
@@ -110,7 +113,7 @@ set(hiopLinAlg_CUDASPARSEKERNELS_SRC
   MatrixSparseCsrCudaKernels.cu
 )
 set(hiopLinAlg_FGMRESKERNELS_SRC
-  KrylovSolverKernels.cu
+  #KrylovSolverKernels.cu
 )
 set(hiopLinAlg_CUDASPARSE_SRC
   hiopMatrixSparseCsrCuda.cpp
@@ -158,8 +161,8 @@ if(HIOP_SPARSE)
     if(HIOP_USE_CUDA)
       list(APPEND hiopLinAlg_SRC ${hiopLinAlg_CUSOLVER_CHOL_SRC})
       set_source_files_properties(${hiopLinAlg_CUSOLVER_CHOL_SRC} PROPERTIES LANGUAGE CUDA)
-      set_source_files_properties(${hiopLinAlg_FGMRESKERNELS_SRC} PROPERTIES LANGUAGE CUDA)
-      list(APPEND hiopLinAlg_SRC ${hiopLinAlg_FGMRESKERNELS_SRC})
+      # set_source_files_properties(${hiopLinAlg_FGMRESKERNELS_SRC} PROPERTIES LANGUAGE CUDA)
+      # list(APPEND hiopLinAlg_SRC ${hiopLinAlg_FGMRESKERNELS_SRC})
     endif(HIOP_USE_CUDA)
     
     if(HIOP_USE_GINKGO)
@@ -224,3 +227,7 @@ install(
 
 add_library(hiopLinAlg OBJECT ${hiopLinAlg_SRC})
 target_link_libraries(hiopLinAlg PRIVATE hiop_tpl)
+# if(HIOP_USE_CUSOLVER_LU)
+#   add_subdirectory(ReSolve)
+#   target_link_libraries(hiopLinAlg PUBLIC ReSolve)
+# endif()
