+++ src/LinAlg/IterativeRefinement.cpp
@@ -786,758 +786,4 @@ namespace ReSolve {
   }
 
 
-
-
-
-  MatrixCsr::MatrixCsr()
-  {
-  }
-
-  MatrixCsr::~MatrixCsr()
-  {
-    if(n_ == 0)
-      return;
-
-    cudaFree(irows_);
-    cudaFree(jcols_);
-    cudaFree(vals_);
-
-    delete [] irows_host_;
-    delete [] jcols_host_;
-    delete [] vals_host_ ;
-  }
-
-  void MatrixCsr::allocate_size(int n)
-  {
-    n_ = n;
-    checkCudaErrors(cudaMalloc(&irows_, (n_+1) * sizeof(int)));
-    irows_host_ = new int[n_+1]{0};
-  }
-
-  void MatrixCsr::allocate_nnz(int nnz)
-  {
-    nnz_ = nnz;
-    checkCudaErrors(cudaMalloc(&jcols_, nnz_ * sizeof(int)));
-    checkCudaErrors(cudaMalloc(&vals_,  nnz_ * sizeof(double)));
-    jcols_host_ = new int[nnz_]{0};
-    vals_host_  = new double[nnz_]{0};
-  }
-
-  void MatrixCsr::update_from_host_mirror()
-  {
-    checkCudaErrors(cudaMemcpy(irows_, irows_host_, sizeof(int)    * (n_+1), cudaMemcpyHostToDevice));
-    checkCudaErrors(cudaMemcpy(jcols_, jcols_host_, sizeof(int)    * nnz_,   cudaMemcpyHostToDevice));
-    checkCudaErrors(cudaMemcpy(vals_,  vals_host_,  sizeof(double) * nnz_,   cudaMemcpyHostToDevice));
-  }
-
-  void MatrixCsr::copy_to_host_mirror()
-  {
-    checkCudaErrors(cudaMemcpy(irows_host_, irows_, sizeof(int)    * (n_+1), cudaMemcpyDeviceToHost));
-    checkCudaErrors(cudaMemcpy(jcols_host_, jcols_, sizeof(int)    * nnz_,   cudaMemcpyDeviceToHost));
-    checkCudaErrors(cudaMemcpy(vals_host_,  vals_,  sizeof(double) * nnz_,   cudaMemcpyDeviceToHost));
-  }
-
-  // Error checking utility for CUDA
-  // KS: might later become part of src/Utils, putting it here for now
-  template <typename T>
-  void MatrixCsr::resolveCheckCudaError(T result,
-                                                  const char* const file,
-                                                  int const line)
-  {
-    if(result) {
-      fprintf(stdout, 
-              "CUDA error at %s:%d, error# %d\n", 
-              file, 
-              line, 
-              result);
-      assert(false);
-    }
-  }
-
-
-
-
-
-  RefactorizationSolver::RefactorizationSolver()
-  {
-    mat_A_csr_ = new MatrixCsr();
-
-    // handles
-    cusparseCreate(&handle_);
-    cusolverSpCreate(&handle_cusolver_);
-    cublasCreate(&handle_cublas_);
-
-    // descriptors
-    cusparseCreateMatDescr(&descr_A_);
-    cusparseSetMatType(descr_A_, CUSPARSE_MATRIX_TYPE_GENERAL);
-    cusparseSetMatIndexBase(descr_A_, CUSPARSE_INDEX_BASE_ZERO);
-  }
-
-  RefactorizationSolver::RefactorizationSolver(int n)
-    : n_(n)
-  {
-    mat_A_csr_ = new MatrixCsr();
-
-    // handles
-    cusparseCreate(&handle_);
-    cusolverSpCreate(&handle_cusolver_);
-    cublasCreate(&handle_cublas_);
-
-    // descriptors
-    cusparseCreateMatDescr(&descr_A_);
-    cusparseSetMatType(descr_A_, CUSPARSE_MATRIX_TYPE_GENERAL);
-    cusparseSetMatIndexBase(descr_A_, CUSPARSE_INDEX_BASE_ZERO);
-
-  }
-
-  RefactorizationSolver::~RefactorizationSolver()
-  {
-    delete mat_A_csr_;
-    if(iterative_refinement_enabled_)
-      delete ir_;
-
-    // Delete workspaces and handles
-    cudaFree(d_work_);
-    cusparseDestroy(handle_);
-    cusolverSpDestroy(handle_cusolver_);
-    cublasDestroy(handle_cublas_);
-    cusparseDestroyMatDescr(descr_A_);
-
-    // Delete residual and solution vectors
-    cudaFree(devr_);
-    cudaFree(devx_);
-
-    // Delete matrix descriptor used in cuSolverGLU setup
-    if(cusolver_glu_enabled_) {
-      cusparseDestroyMatDescr(descr_M_);
-      cusolverSpDestroyGluInfo(info_M_);
-    }
-
-    if(cusolver_rf_enabled_) {
-      cudaFree(d_P_);
-      cudaFree(d_Q_);
-      cudaFree(d_T_);
-    }
-
-    klu_free_symbolic(&Symbolic_, &Common_);
-    klu_free_numeric(&Numeric_, &Common_);
-    delete [] mia_;
-    delete [] mja_;
-  }
-
-  void RefactorizationSolver::enable_iterative_refinement()
-  {
-    ir_ = new IterativeRefinement();
-    if(ir_ != nullptr)
-      iterative_refinement_enabled_ = true;
-  }
-
-  // TODO: Refactor to only pass mat_A_csr_ to setup_system_matrix; n and nnz can be read from mat_A_csr_
-  void RefactorizationSolver::setup_iterative_refinement_matrix(int n, int nnz)
-  {
-    ir_->setup_system_matrix(n, nnz, mat_A_csr_->get_irows(), mat_A_csr_->get_jcols(), mat_A_csr_->get_vals());
-  }
-
-  // TODO: Can this function be merged with setup_iterative_refinement_matrix ?
-  void RefactorizationSolver::configure_iterative_refinement(cusparseHandle_t   cusparse_handle,
-                                                             cublasHandle_t     cublas_handle,
-                                                             cusolverRfHandle_t cusolverrf_handle,
-                                                             int n,
-                                                             double* d_T,
-                                                             int* d_P,
-                                                             int* d_Q,
-                                                             double* devx,
-                                                             double* devr)
-  {
-    ir_->setup(cusparse_handle, cublas_handle, cusolverrf_handle, n, d_T, d_P, d_Q, devx, devr);
-  }
-
-
-  int RefactorizationSolver::setup_factorization()
-  {
-    int* row_ptr = mat_A_csr_->get_irows_host();
-    int* col_idx = mat_A_csr_->get_jcols_host();
-
-    if(fact_ == "klu") {
-      /* initialize KLU setup parameters, dont factorize yet */
-      initializeKLU();
-
-      /*perform KLU but only the symbolic analysis (important)   */
-      klu_free_symbolic(&Symbolic_, &Common_);
-      klu_free_numeric(&Numeric_, &Common_);
-      Symbolic_ = klu_analyze(n_, row_ptr, col_idx, &Common_);
-
-      if(Symbolic_ == nullptr) {
-        return -1;
-      }
-    } else { // for future
-      assert(0 && "Only KLU is available for the first factorization.\n");
-    }
-    return 0;
-  }
-
-  int RefactorizationSolver::factorize()
-  {
-    Numeric_ = klu_factor(mat_A_csr_->get_irows_host(), mat_A_csr_->get_jcols_host(), mat_A_csr_->get_vals_host(), Symbolic_, &Common_);
-    return (Numeric_ == nullptr) ? -1 : 0;
-  }
-
-  void RefactorizationSolver::setup_refactorization()
-  {
-    if(refact_ == "glu") {
-      initializeCusolverGLU();
-      refactorizationSetupCusolverGLU();
-    } else if(refact_ == "rf") {
-      initializeCusolverRf();
-      refactorizationSetupCusolverRf();
-      if(use_ir_ == "yes") {
-        configure_iterative_refinement(handle_, handle_cublas_, handle_rf_, n_, d_T_, d_P_, d_Q_, devx_, devr_);
-      }
-    } else { // for future -
-      assert(0 && "Only glu and rf refactorizations available.\n");
-    }
-  }
-
-  int RefactorizationSolver::refactorize()
-  {
-    // TODO: This memcpy should not be in this function.
-    checkCudaErrors(cudaMemcpy(mat_A_csr_->get_vals(), mat_A_csr_->get_vals_host(), sizeof(double) * nnz_, cudaMemcpyHostToDevice));
-
-    if(refact_ == "glu") {
-      sp_status_ = cusolverSpDgluReset(handle_cusolver_, 
-                                        n_,
-                                        /* A is original matrix */
-                                        nnz_,
-                                        descr_A_,
-                                        mat_A_csr_->get_vals(),  //da_,
-                                        mat_A_csr_->get_irows(), //dia_,
-                                        mat_A_csr_->get_jcols(), //dja_,
-                                        info_M_);
-      sp_status_ = cusolverSpDgluFactor(handle_cusolver_, info_M_, d_work_);
-    } else {
-      if(refact_ == "rf") {
-        sp_status_ = cusolverRfResetValues(n_, 
-                                            nnz_, 
-                                            mat_A_csr_->get_irows(), //dia_,
-                                            mat_A_csr_->get_jcols(), //dja_,
-                                            mat_A_csr_->get_vals(),  //da_,
-                                            d_P_,
-                                            d_Q_,
-                                            handle_rf_);
-        cudaDeviceSynchronize();
-        sp_status_ = cusolverRfRefactor(handle_rf_);
-      }
-    }
-    return 0;
-  }
-
-  bool RefactorizationSolver::triangular_solve(double* dx, const double* rhs, double tol)
-  {
-    if(refact_ == "glu") {
-      sp_status_ = cusolverSpDgluSolve(handle_cusolver_,
-                                       n_,
-                                       /* A is original matrix */
-                                       nnz_,
-                                       descr_A_,
-                                       mat_A_csr_->get_vals(),  //da_,
-                                       mat_A_csr_->get_irows(), //dia_,
-                                       mat_A_csr_->get_jcols(), //dja_,
-                                       devr_,/* right hand side */
-                                       devx_,/* left hand side */
-                                       &ite_refine_succ_,
-                                       &r_nrminf_,
-                                       info_M_,
-                                       d_work_);
-
-      if(sp_status_ == 0) {
-        checkCudaErrors(cudaMemcpy(dx, devx_, sizeof(double) * n_, cudaMemcpyDeviceToHost));
-      } else {
-        // TODO: Implement ReSolve logging system to output these messages
-        std::cout << "GLU solve failed with status: " << sp_status_ << "\n";
-        return false;
-      }
-    } else {
-      if(refact_ == "rf") {
-        if(!is_first_solve_) {
-          sp_status_ = cusolverRfSolve(handle_rf_,
-                                       d_P_,
-                                       d_Q_,
-                                       1,
-                                       d_T_,
-                                       n_,
-                                       devr_,
-                                       n_);
-
-          if(sp_status_ == 0) {
-            if(use_ir_ == "yes") {
-              // Set tolerance based on barrier parameter mu
-              ir_->set_tol(tol);
-              // TODO: Implement ReSolve logging system to output these messages
-              // nlp_->log->printf(hovScalars,
-              //                   "Running iterative refinement with tol %e\n", tol);
-              checkCudaErrors(cudaMemcpy(devx_, rhs, sizeof(double) * n_, cudaMemcpyHostToDevice));
-
-              ir_->fgmres(devr_, devx_);
-              // TODO: Implement ReSolve logging system to output these messages
-              // nlp_->log->printf(hovScalars, 
-              //                   "\t fgmres: init residual norm  %e final residual norm %e number of iterations %d\n", 
-              //                   ir_->getInitialResidalNorm(), 
-              //                   ir_->getFinalResidalNorm(), 
-              //                   ir_->getFinalNumberOfIterations());
-            }
-            checkCudaErrors(cudaMemcpy(dx, devr_, sizeof(double) * n_, cudaMemcpyDeviceToHost));
-
-
-          } else {
-            // TODO: Implement ReSolve logging system to output these messages
-            std::cout << "Rf solve failed with status: " << sp_status_ << "\n";
-            return false;
-          }
-        } else {
-          memcpy(dx, rhs, sizeof(double) * n_);
-          int ok = klu_solve(Symbolic_, Numeric_, n_, 1, dx, &Common_);
-          klu_free_numeric(&Numeric_, &Common_);
-          klu_free_symbolic(&Symbolic_, &Common_);
-          is_first_solve_ = false;
-        }
-      } else {
-        // TODO: Implement ReSolve logging system to output these messages
-        std::cout << "Unknown refactorization, exiting\n";
-        assert(false && "Only GLU and cuSolverRf are available refactorizations.");
-      }
-    }
-    return true;
-  }
-
-  // helper private function needed for format conversion
-  int RefactorizationSolver::createM(const int n, 
-                                     const int /* nnzL */,
-                                     const int* Lp, 
-                                     const int* Li,
-                                     const int /* nnzU */, 
-                                     const int* Up,
-                                     const int* Ui)
-  {
-    int row;
-    for(int i = 0; i < n; ++i) {
-      // go through EACH COLUMN OF L first
-      for(int j = Lp[i]; j < Lp[i + 1]; ++j) {
-        row = Li[j];
-        // BUT dont count diagonal twice, important
-        if(row != i) {
-          mia_[row + 1]++;
-        }
-      }
-      // then each column of U
-      for(int j = Up[i]; j < Up[i + 1]; ++j) {
-        row = Ui[j];
-        mia_[row + 1]++;
-      }
-    }
-    // then organize mia_;
-    mia_[0] = 0;
-    for(int i = 1; i < n + 1; i++) {
-      mia_[i] += mia_[i - 1];
-    }
-
-    std::vector<int> Mshifts(n, 0);
-    for(int i = 0; i < n; ++i) {
-      // go through EACH COLUMN OF L first
-      for(int j = Lp[i]; j < Lp[i + 1]; ++j) {
-        row = Li[j];
-        if(row != i) {
-          // place (row, i) where it belongs!
-          mja_[mia_[row] + Mshifts[row]] = i;
-          Mshifts[row]++;
-        }
-      }
-      // each column of U next
-      for(int j = Up[i]; j < Up[i + 1]; ++j) {
-        row = Ui[j];
-        mja_[mia_[row] + Mshifts[row]] = i;
-        Mshifts[row]++;
-      }
-    }
-    return 0;
-  }
-
-  int RefactorizationSolver::initializeKLU()
-  {
-    klu_defaults(&Common_);
-
-    // TODO: consider making this a part of setup options so that user can
-    // set up these values. For now, we keep them hard-wired.
-    Common_.btf = 0;
-    Common_.ordering = ordering_; // COLAMD=1; AMD=0
-    Common_.tol = 0.1;
-    Common_.scale = -1;
-    Common_.halt_if_singular = 1;
-
-    return 0;
-  }
-
-  int RefactorizationSolver::initializeCusolverGLU()
-  {
-    // nlp_->log->printf(hovScalars, "CUSOLVER: Glu \n");
-    cusparseCreateMatDescr(&descr_M_);
-    cusparseSetMatType(descr_M_, CUSPARSE_MATRIX_TYPE_GENERAL);
-    cusparseSetMatIndexBase(descr_M_, CUSPARSE_INDEX_BASE_ZERO);
-
-    // info (data structure where factorization is stored)
-    // this is done in the constructor - however, this function might be called more than once
-    cusolverSpDestroyGluInfo(info_M_);
-    cusolverSpCreateGluInfo(&info_M_);
-
-    cusolver_glu_enabled_ = true;
-    return 0;
-  }
-
-  int RefactorizationSolver::initializeCusolverRf()
-  {
-    // nlp_->log->printf(hovScalars, "CUSOLVER: Rf \n");
-    cusolverRfCreate(&handle_rf_);
-
-    checkCudaErrors(cusolverRfSetAlgs(handle_rf_,
-                                      CUSOLVERRF_FACTORIZATION_ALG2,
-                                      CUSOLVERRF_TRIANGULAR_SOLVE_ALG2));
-
-    checkCudaErrors(cusolverRfSetMatrixFormat(handle_rf_, 
-                                              CUSOLVERRF_MATRIX_FORMAT_CSR,
-                                              CUSOLVERRF_UNIT_DIAGONAL_STORED_L));
-
-    cusolverRfSetResetValuesFastMode(handle_rf_,
-                                     CUSOLVERRF_RESET_VALUES_FAST_MODE_ON);
-
-    const double boost = 1e-12;
-    const double zero = 1e-14;
-
-    cusolverRfSetNumericProperties(handle_rf_, zero, boost);
-
-    cusolver_rf_enabled_ = true;
-    return 0;
-  }
-
-  // call if both the matrix and the nnz structure changed or if convergence is
-  // poor while using refactorization.
-  int RefactorizationSolver::refactorizationSetupCusolverGLU()
-  {
-    // for now this ONLY WORKS if proceeded by KLU. Might be worth decoupling
-    // later
-
-    // get sizes
-    const int nnzL = Numeric_->lnz;
-    const int nnzU = Numeric_->unz;
-
-    const int nnzM = (nnzL + nnzU - n_);
-
-    /* parse the factorization */
-
-    mia_ = new int[n_ + 1]{0};
-    mja_ = new int[nnzM]{0};
-    int* Lp = new int[n_ + 1];
-    int* Li = new int[nnzL];
-    // we can't use nullptr instead od Lx and Ux because it causes SEG FAULT. It
-    // seems like a waste of memory though.
-
-    double* Lx = new double[nnzL];
-    int* Up = new int[n_ + 1];
-    int* Ui = new int[nnzU];
-
-    double* Ux = new double[nnzU];
-
-    int ok = klu_extract(Numeric_, 
-                         Symbolic_, 
-                         Lp, 
-                         Li, 
-                         Lx, 
-                         Up, 
-                         Ui, 
-                         Ux, 
-                         nullptr,
-                         nullptr, 
-                         nullptr, 
-                         nullptr, 
-                         nullptr, 
-                         nullptr, 
-                         nullptr,
-                         &Common_);
-    createM(n_, nnzL, Lp, Li, nnzU, Up, Ui);
-
-    delete[] Lp;
-    delete[] Li;
-    delete[] Lx;
-    delete[] Up;
-    delete[] Ui;
-    delete[] Ux;
-
-    /* setup GLU */
-    sp_status_ = cusolverSpDgluSetup(handle_cusolver_, 
-                                     n_,
-                                     nnz_, 
-                                     descr_A_, 
-                                     mat_A_csr_->get_irows_host(), //kRowPtr_,
-                                     mat_A_csr_->get_jcols_host(), //jCol_, 
-                                     Numeric_->Pnum, /* base-0 */
-                                     Symbolic_->Q,   /* base-0 */
-                                     nnzM,           /* nnzM */
-                                     descr_M_, 
-                                     mia_, 
-                                     mja_, 
-                                     info_M_);
-
-    sp_status_ = cusolverSpDgluBufferSize(handle_cusolver_, info_M_, &size_M_);
-    assert(CUSOLVER_STATUS_SUCCESS == sp_status_);
-
-    buffer_size_ = size_M_;
-    checkCudaErrors(cudaMalloc((void**)&d_work_, buffer_size_));
-
-    sp_status_ = cusolverSpDgluAnalysis(handle_cusolver_, info_M_, d_work_);
-    assert(CUSOLVER_STATUS_SUCCESS == sp_status_);
-
-    // reset and refactor so factors are ON THE GPU
-
-    sp_status_ = cusolverSpDgluReset(handle_cusolver_, 
-                                     n_,
-                                     /* A is original matrix */
-                                     nnz_, 
-                                     descr_A_, 
-                                     mat_A_csr_->get_vals(),  //da_, 
-                                     mat_A_csr_->get_irows(), //dia_, 
-                                     mat_A_csr_->get_jcols(), //dja_, 
-                                     info_M_);
-
-    assert(CUSOLVER_STATUS_SUCCESS == sp_status_);
-    sp_status_ = cusolverSpDgluFactor(handle_cusolver_, info_M_, d_work_);
-    return 0;
-  }
-
-  int RefactorizationSolver::refactorizationSetupCusolverRf()
-  {
-    // for now this ONLY WORKS if preceeded by KLU. Might be worth decoupling
-    // later
-    const int nnzL = Numeric_->lnz;
-    const int nnzU = Numeric_->unz;
-
-    checkCudaErrors(cudaMalloc(&d_P_, (n_) * sizeof(int)));
-    checkCudaErrors(cudaMalloc(&d_Q_, (n_) * sizeof(int)));
-    checkCudaErrors(cudaMalloc(&d_T_, (n_) * sizeof(double)));
-
-    checkCudaErrors(cudaMemcpy(d_P_, Numeric_->Pnum, sizeof(int) * (n_), cudaMemcpyHostToDevice));
-    checkCudaErrors(cudaMemcpy(d_Q_, Symbolic_->Q, sizeof(int) * (n_), cudaMemcpyHostToDevice));
-
-    int* Lp = new int[n_ + 1];
-    int* Li = new int[nnzL];
-    double* Lx = new double[nnzL];
-    int* Up = new int[n_ + 1];
-    int* Ui = new int[nnzU];
-    double* Ux = new double[nnzU];
-
-    int ok = klu_extract(Numeric_, 
-                         Symbolic_, 
-                         Lp, 
-                         Li, 
-                         Lx, 
-                         Up, 
-                         Ui, 
-                         Ux, 
-                         nullptr, 
-                         nullptr, 
-                         nullptr, 
-                         nullptr, 
-                         nullptr,
-                         nullptr,
-                         nullptr,
-                         &Common_);
-
-    /* CSC */
-    int* d_Lp;
-    int* d_Li;
-    int* d_Up;
-    int* d_Ui;
-    double* d_Lx;
-    double* d_Ux;
-    /* CSR */
-    int* d_Lp_csr;
-    int* d_Li_csr;
-    int* d_Up_csr;
-    int* d_Ui_csr;
-    double* d_Lx_csr;
-    double* d_Ux_csr;
-
-    /* allocate CSC */
-    checkCudaErrors(cudaMalloc(&d_Lp, (n_ + 1) * sizeof(int)));
-    checkCudaErrors(cudaMalloc(&d_Li, nnzL * sizeof(int)));
-    checkCudaErrors(cudaMalloc(&d_Lx, nnzL * sizeof(double)));
-    checkCudaErrors(cudaMalloc(&d_Up, (n_ + 1) * sizeof(int)));
-    checkCudaErrors(cudaMalloc(&d_Ui, nnzU * sizeof(int)));
-    checkCudaErrors(cudaMalloc(&d_Ux, nnzU * sizeof(double)));
-
-    /* allocate CSR */
-    checkCudaErrors(cudaMalloc(&d_Lp_csr, (n_ + 1) * sizeof(int)));
-    checkCudaErrors(cudaMalloc(&d_Li_csr, nnzL * sizeof(int)));
-    checkCudaErrors(cudaMalloc(&d_Lx_csr, nnzL * sizeof(double)));
-    checkCudaErrors(cudaMalloc(&d_Up_csr, (n_ + 1) * sizeof(int)));
-    checkCudaErrors(cudaMalloc(&d_Ui_csr, nnzU * sizeof(int)));
-    checkCudaErrors(cudaMalloc(&d_Ux_csr, nnzU * sizeof(double)));
-
-    /* copy CSC to the GPU */
-    checkCudaErrors(cudaMemcpy(d_Lp, Lp, sizeof(int) * (n_ + 1), cudaMemcpyHostToDevice));
-    checkCudaErrors(cudaMemcpy(d_Li, Li, sizeof(int) * (nnzL), cudaMemcpyHostToDevice));
-    checkCudaErrors(cudaMemcpy(d_Lx, Lx, sizeof(double) * (nnzL), cudaMemcpyHostToDevice));
-
-    checkCudaErrors(cudaMemcpy(d_Up, Up, sizeof(int) * (n_ + 1), cudaMemcpyHostToDevice));
-    checkCudaErrors(cudaMemcpy(d_Ui, Ui, sizeof(int) * (nnzU), cudaMemcpyHostToDevice));
-    checkCudaErrors(cudaMemcpy(d_Ux, Ux, sizeof(double) * (nnzU), cudaMemcpyHostToDevice));
-
-    /* we dont need these any more */
-    delete[] Lp;
-    delete[] Li;
-    delete[] Lx;
-    delete[] Up;
-    delete[] Ui;
-    delete[] Ux;
-
-    /* now CSC to CSR using the new cuda 11 awkward way */
-    size_t bufferSizeL;
-    size_t bufferSizeU;
-
-    cusparseStatus_t csp = cusparseCsr2cscEx2_bufferSize(handle_, 
-                                                         n_, 
-                                                         n_, 
-                                                         nnzL, 
-                                                         d_Lx, 
-                                                         d_Lp, 
-                                                         d_Li, 
-                                                         d_Lx_csr, 
-                                                         d_Lp_csr,
-                                                         d_Li_csr, 
-                                                         CUDA_R_64F, 
-                                                         CUSPARSE_ACTION_NUMERIC,
-                                                         CUSPARSE_INDEX_BASE_ZERO, 
-                                                         CUSPARSE_CSR2CSC_ALG1, 
-                                                         &bufferSizeL);
-
-    csp = cusparseCsr2cscEx2_bufferSize(handle_, 
-                                        n_, 
-                                        n_, 
-                                        nnzU, 
-                                        d_Ux, 
-                                        d_Up, 
-                                        d_Ui, 
-                                        d_Ux_csr, 
-                                        d_Up_csr, 
-                                        d_Ui_csr, 
-                                        CUDA_R_64F,
-                                        CUSPARSE_ACTION_NUMERIC,
-                                        CUSPARSE_INDEX_BASE_ZERO,
-                                        CUSPARSE_CSR2CSC_ALG1,
-                                        &bufferSizeU);
-    /* allocate buffers */
-
-    double* d_workL;
-    double* d_workU;
-    checkCudaErrors(cudaMalloc((void**)&d_workL, bufferSizeL));
-    checkCudaErrors(cudaMalloc((void**)&d_workU, bufferSizeU));
-
-    /* actual CSC to CSR */
-
-    csp = cusparseCsr2cscEx2(handle_, 
-                             n_, 
-                             n_, 
-                             nnzL, 
-                             d_Lx, 
-                             d_Lp, 
-                             d_Li,
-                             d_Lx_csr, 
-                             d_Lp_csr,
-                             d_Li_csr,
-                             CUDA_R_64F,
-                             CUSPARSE_ACTION_NUMERIC,
-                             CUSPARSE_INDEX_BASE_ZERO,
-                             CUSPARSE_CSR2CSC_ALG1,
-                             d_workL);
-
-    csp = cusparseCsr2cscEx2(handle_,
-                             n_,
-                             n_,
-                             nnzU,
-                             d_Ux, 
-                             d_Up, 
-                             d_Ui, 
-                             d_Ux_csr, 
-                             d_Up_csr, 
-                             d_Ui_csr, 
-                             CUDA_R_64F,
-                             CUSPARSE_ACTION_NUMERIC,
-                             CUSPARSE_INDEX_BASE_ZERO,
-                             CUSPARSE_CSR2CSC_ALG1,
-                             d_workU);
-
-    (void)csp; // mute unused variable warnings
-
-    /* CSC no longer needed, nor the work arrays! */
-
-    cudaFree(d_Lp);
-    cudaFree(d_Li);
-    cudaFree(d_Lx);
-
-    cudaFree(d_Up);
-    cudaFree(d_Ui);
-    cudaFree(d_Ux);
-
-    cudaFree(d_workU);
-    cudaFree(d_workL);
-
-    /* actual setup */
-
-    sp_status_ = cusolverRfSetupDevice(n_, 
-                                       nnz_,
-                                       mat_A_csr_->get_irows(), //dia_,
-                                       mat_A_csr_->get_jcols(), //dja_,
-                                       mat_A_csr_->get_vals(),  //da_,
-                                       nnzL,
-                                       d_Lp_csr,
-                                       d_Li_csr,
-                                       d_Lx_csr,
-                                       nnzU,
-                                       d_Up_csr,
-                                       d_Ui_csr,
-                                       d_Ux_csr,
-                                       d_P_,
-                                       d_Q_,
-                                       handle_rf_);
-    cudaDeviceSynchronize();
-    sp_status_ = cusolverRfAnalyze(handle_rf_);
-
-    //clean up 
-    cudaFree(d_Lp_csr);
-    cudaFree(d_Li_csr);
-    cudaFree(d_Lx_csr);
-
-    cudaFree(d_Up_csr);
-    cudaFree(d_Ui_csr);
-    cudaFree(d_Ux_csr);
-
-    return 0;
-  }
-
-
-
-  // Error checking utility for CUDA
-  // KS: might later become part of src/Utils, putting it here for now
-  template <typename T>
-  void RefactorizationSolver::resolveCheckCudaError(T result,
-                                                  const char* const file,
-                                                  int const line)
-  {
-    if(result) {
-      fprintf(stdout, 
-              "CUDA error at %s:%d, error# %d\n", 
-              file, 
-              line, 
-              result);
-      assert(false);
-    }
-  }
-
 } // namespace ReSolve
